Division del codigo en varios componentes y nuevos elementos añadidos
