+++ src/front/components/QuestionCard.css
@@ -1,0 +1,80 @@
+.question-card {
+  background-color: var(--dark-color);
+  max-width: 700px;
+  width: 90%;
+  padding: 1.5rem;
+  border-radius: 0.5rem;
+  box-shadow: 0 0.5rem 1rem rgba(0, 0, 0, 0.15);
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+.timer-display {
+  color: var(--text-light);
+  font-size: 1.1rem;
+  font-weight: bold;
+  margin-bottom: 1.5rem;
+  display: flex;
+  align-items: center;
+  gap: 0.5rem;
+}
+
+.timer-number {
+  font-size: 1.4rem;
+  color: var(--accent-color);
+  min-width: 30px;
+  text-align: right;
+}
+
+.timer-warning {
+  color: var(--danger-color);
+  animation: pulse 1s infinite alternate;
+}
+
+@keyframes pulse {
+  from { transform: scale(1); opacity: 1; }
+  to { transform: scale(1.05); opacity: 0.8; }
+}
+
+.question-text {
+  color: var(--text-light);
+  margin-bottom: 3rem;
+  text-align: center;
+}
+
+.answers-grid {
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  gap: 1rem;
+}
+
+.answer-button {
+  background-color: var(--accent-color);
+  border-color: var(--accent-color);
+  color: var(--text-dark);
+  transition: background-color 0.4s ease, border-color 0.4s ease, color 0.4s ease;
+}
+
+.answer-button:hover:not(:disabled) {
+  background-color: #c9a75c;
+  border-color: #c9a75c;
+}
+
+.btn-success-feedback {
+  background-color: var(--success-color) !important;
+  border-color: var(--success-color) !important;
+  color: var(--text-light) !important;
+}
+
+.btn-danger-feedback {
+  background-color: var(--danger-color) !important;
+  border-color: var(--danger-color) !important;
+  color: var(--text-light) !important;
+}
+
+.answer-button:disabled {
+  opacity: 0.6;
+  cursor: not-allowed;
+}
