+++ src/front/data/questions.js
@@ -1,0 +1,57 @@
+export const allQuestions = [
+  {
+    question: '¿Cuál es la capital de Francia?',
+    answers: ['Londres', 'Berlín', 'París', 'Roma'],
+    correctAnswer: 'París',
+  },
+  {
+    question: '¿Qué planeta es conocido como el Planeta Rojo?',
+    answers: ['Júpiter', 'Marte', 'Venus', 'Saturno'],
+    correctAnswer: 'Marte',
+  },
+  {
+    question: '¿Cuántos continentes hay en la Tierra?',
+    answers: ['5', '6', '7', '8'],
+    correctAnswer: '7',
+  },
+  {
+    question: '¿Quién escribió "El Quijote"?',
+    answers: ['Miguel de Cervantes', 'Gabriel García Márquez', 'Federico García Lorca', 'Jorge Luis Borges'],
+    correctAnswer: 'Miguel de Cervantes',
+  },
+  {
+    question: '¿En qué año llegó el hombre a la luna?',
+    answers: ['1965', '1969', '1971', '1975'],
+    correctAnswer: '1969',
+  },
+  {
+    question: '¿Cuál es el océano más grande de la Tierra?',
+    answers: ['Atlántico', 'Índico', 'Ártico', 'Pacífico'],
+    correctAnswer: 'Pacífico',
+  },
+  {
+    question: '¿Qué elemento químico tiene el símbolo "O"?',
+    answers: ['Oro', 'Oxígeno', 'Osmio', 'Óxido'],
+    correctAnswer: 'Oxígeno',
+  },
+  {
+    question: '¿Cuál es la moneda de Japón?',
+    answers: ['Yuan', 'Won', 'Yen', 'Dólar'],
+    correctAnswer: 'Yen',
+  },
+  {
+    question: '¿Cuántos huesos tiene el cuerpo humano adulto?',
+    answers: ['206', '210', '198', '200'],
+    correctAnswer: '206',
+  },
+  {
+    question: '¿Qué famoso científico formuló la Teoría de la Relatividad?',
+    answers: ['Isaac Newton', 'Galileo Galilei', 'Albert Einstein', 'Stephen Hawking'],
+    correctAnswer: 'Albert Einstein',
+  },
+  {
+    question: '¿En qué país se encuentra la Gran Muralla China?',
+    answers: ['India', 'Japón', 'China', 'Corea del Sur'],
+    correctAnswer: 'China',
+  },
+];
