--- src/front/pages/SalaDeJuegos.css
@@ -1,117 +1,0 @@
-:root {
-  --background-color: #5ea9a7;
-  --dark-color: #23322d;
-  --accent-color: #d6b770;
-  --purple-color: #9170CB;
-  --text-light: white;
-  --text-dark: #23322d;
-  --success-color: green;
-  --danger-color: red;
-}
-
-.game-container {
-  background-color: var(--background-color);
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-}
-
-.game-header {
-  height: 100px;
-  padding: 1rem;
-  display: flex;
-  justify-content: flex-start;
-  align-items: flex-start;
-}
-
-.user-profile-box {
-  background-color: var(--dark-color);
-  color: var(--text-light);
-  width: 220px;
-  padding: 1rem;
-  border-radius: 0.5rem;
-  box-shadow: 0 0.125rem 0.25rem rgba(0, 0, 0, 0.075);
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  margin-left: 20px;
-}
-
-.user-avatar-emoji {
-  font-size: 2em;
-  margin-right: 0.5rem;
-}
-
-.user-name {
-  color: var(--text-light);
-  font-weight: bold;
-  font-size: 1.25rem;
-}
-
-.score-star {
-  font-size: 1.5em;
-  margin-right: 0.25rem;
-}
-
-.score-number {
-  color: var(--accent-color);
-  font-weight: bold;
-  font-size: 1.25rem;
-}
-
-.game-main {
-  flex-grow: 1;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  padding: 1rem;
-}
-
-.question-card {
-  background-color: var(--dark-color);
-  max-width: 700px;
-  width: 90%;
-  padding: 1.5rem;
-  border-radius: 0.5rem;
-  box-shadow: 0 0.5rem 1rem rgba(0, 0, 0, 0.15);
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-.question-text {
-  color: var(--text-light);
-  margin-bottom: 3rem;
-  text-align: center;
-}
-
-.answers-grid {
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  gap: 1rem;
-}
-
-.answer-button {
-  background-color: var(--accent-color);
-  border-color: var(--accent-color);
-  color: var(--text-dark);
-  transition: background-color 0.4s ease, border-color 0.4s ease, color 0.4s ease;
-}
-
-.answer-button:hover {
-  background-color: #c9a75c;
-  border-color: #c9a75c;
-}
-
-.btn-success-feedback {
-  background-color: var(--success-color) !important;
-  border-color: var(--success-color) !important;
-  color: var(--text-light) !important;
-}
-
-.btn-danger-feedback {
-  background-color: var(--danger-color) !important;
-  border-color: var(--danger-color) !important;
-  color: var(--text-light) !important;
-}
