+++ src/front/pages/davidPages/SalaDeJuegos.css
@@ -1,0 +1,27 @@
+:root {
+  --background-color: #5ea9a7;
+  --dark-color: #23322d;
+  --accent-color: #d6b770;
+  --purple-color: #9170CB;
+  --text-light: white;
+  --text-dark: #23322d;
+  --success-color: green;
+  --danger-color: red;
+}
+
+.game-container {
+  background-color: var(--background-color);
+  min-height: 100vh;
+  display: flex;
+  flex-direction: column;
+}
+
+.game-header {
+  height: 100px;
+  padding: 1rem;
+  display: flex;
+  justify-content: space-between; 
+  align-items: flex-start;
+  position: relative; 
+}
+
